requirements.txt
aiogram==2.16
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.5
amqp==5.0.6
async-timeout==4.0.1
attrs==21.2.0
autoflake==1.4
Babel==2.9.1
billiard==3.6.4.0
black==21.11b0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
flake8==4.0.1
frozenlist==1.2.0
greenlet==1.1.2
idna==3.3
invoke==1.6.0
isort==5.10.1
kombu==5.2.2
Mako==1.1.5
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.2.0
mypy==0.910
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
prompt-toolkit==3.0.22
pycodestyle==2.8.0
pyflakes==2.4.0
python-decouple==3.5
pytz==2021.3
redis==4.0.1
regex==2021.11.10
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
yarl==1.7.2
